Se organizan los Service y se agregan pruebas unitarias

+++ src/main/java/com/familyspencesapi/service/task/TaskService.java
@@ -1,74 +1,115 @@
 package com.familyspencesapi.service.task;
 
 import com.familyspencesapi.domain.tasks.Tasks;
-import com.familyspencesapi.mock.tasks.TaskMockData;
+import com.familyspencesapi.repositories.expense.ExpenseRepository;
 import com.familyspencesapi.repositories.task.ITaskRepository;
-import org.springframework.beans.BeanUtils;
+import com.familyspencesapi.repositories.users.FamilyRepository;
+import com.familyspencesapi.repositories.vacation.VacationRepository;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
 @Service
 public class TaskService {
 
-    private ITaskRepository iTaskRepository;
+    private final ITaskRepository iTaskRepository;
+    private final FamilyRepository familyRepository;
+    private final ExpenseRepository expenseRepository;
+    private final VacationRepository vacationRepository;
 
-    private final List<Tasks> tasks;
-
-    public TaskService(ITaskRepository iTaskRepository) {
+    public TaskService(ITaskRepository iTaskRepository, FamilyRepository familyRepository, ExpenseRepository expenseRepository, VacationRepository vacationRepository) {
         this.iTaskRepository = iTaskRepository;
-        this.tasks = TaskMockData.getAllMockTasks();
+        this.familyRepository = familyRepository;
+        this.expenseRepository = expenseRepository;
+        this.vacationRepository = vacationRepository;
     }
 
     public List<Tasks> getAllTasks(final UUID familyId){
 
         List<Tasks> tasksList = iTaskRepository.findByFamilyId(familyId);
 
-        if (tasksList.isEmpty()) {
+        if(tasksList.isEmpty()){
             throw new IllegalArgumentException("Family identification does not exist");
         }
+
         return tasksList;
     }
 
+
     public Tasks getTask(final UUID familyId, final UUID taskId){
 
-        return tasks.stream().filter(t -> taskId.equals(t.getId()) && familyId.equals(t.getFamilyId()))
-                .findFirst().orElseThrow(()-> new IllegalArgumentException("Family or task identification does not exist")) ;
+        if(familyRepository.findById(familyId).isEmpty()){
+            throw new IllegalArgumentException("Family not found");
+        }
+
+        return iTaskRepository.findById(taskId)
+                .orElseThrow(() -> new IllegalArgumentException("Task not found "));
     }
 
+
     public String saveTask(final UUID familyId, final Tasks task){
-        if (tasks.stream().noneMatch(t -> t.getId().equals(task.getId()))) {
-            throw new IllegalArgumentException("Family identification does not exist");
+
+        if (task.getIdVacations() != null && task.getIdExpenseve() != null) {
+            throw new IllegalArgumentException("A task cannot have both Vacation and Expenseve");
         }
 
-        if (tasks.stream().noneMatch(t -> t.getId().equals(task.getId()))) {
-            throw new IllegalArgumentException("Task with same id");
+        if(familyRepository.findById(familyId).isEmpty()){
+            throw new IllegalArgumentException("Family not found");
         }
         iTaskRepository.save(task);
+
         return "Task registered successfully";
     }
 
-    public String updateTask(final UUID familyId, final UUID taskId, final Tasks task){
-        Tasks existingTask = tasks.stream()
-                .filter(t -> t.getFamilyId().equals(familyId) && t.getId().equals(taskId))
-                .findFirst()
-                .orElseThrow(() -> new IllegalArgumentException("Family identification or task id does not exist"));
 
-        BeanUtils.copyProperties(task,existingTask,"id","familyId");
+    public String updateTask(final UUID familyId, final UUID taskId, Tasks task) {
+        if (familyRepository.findById(familyId).isEmpty()) {
+            throw new IllegalArgumentException("Family not found");
+        }
+        Tasks taskUpdate = iTaskRepository.findById(taskId)
+                .orElseThrow(() -> new IllegalArgumentException("Task not found"));
+        taskUpdate.setName(task.getName());
+        taskUpdate.setDescription(task.getDescription());
+        taskUpdate.setStatus(taskUpdate.isStatus());
+
+        if (task.getIdVacations() != null && task.getIdExpenseve() != null) {
+            throw new IllegalArgumentException("A task cannot have both Vacation and Expenseve");
+        }
+        if (task.getIdVacations() != null) {
+            taskUpdate.setIdVacations(
+                    vacationRepository.findById(task.getIdVacations().getId())
+                            .orElseThrow(() -> new IllegalArgumentException("Vacation not found"))
+            );
+            taskUpdate.setIdExpenseve(null);
+        }
+
+        if (task.getIdExpenseve() != null) {
+            taskUpdate.setIdExpenseve(
+                    expenseRepository.findById(task.getIdExpenseve().getId())
+                            .orElseThrow(() -> new IllegalArgumentException("Expenseve not found"))
+            );
+            taskUpdate.setIdVacations(null);
+        }
+        iTaskRepository.save(taskUpdate);
         return "Task updated successfully";
     }
 
-    public String deleteTask(final UUID familyId, final UUID taskId){
 
-        boolean tasksExist = tasks.removeIf(t -> t.getId().equals(taskId) && t.getFamilyId().equals(familyId));
+    public String deleteTask(final UUID familyId, final UUID taskId) {
 
-        iTaskRepository.deleteById(taskId);
+        if(familyRepository.findById(familyId).isEmpty()){
+            throw new IllegalArgumentException("Family not found");
+        }
 
-        if(!tasksExist){
+        Tasks tasksExist = iTaskRepository.findById(taskId)
+                .orElseThrow(() -> new IllegalArgumentException("Task not found"));
+
+        if(tasksExist == null){
             throw new IllegalArgumentException("Family identification or task id does not exist");
         }
+
+        iTaskRepository.deleteById(taskId);
 
         return "Task deleted successfully";
     }
